% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4health.tutorials.R
\docType{package}
\name{r4health.tutorials}
\alias{r4health.tutorials}
\alias{r4health.tutorials-package}
\title{r4health.tutorials}
\description{
This package simply provides tutorials for learning R. It has dependencies on
the packages that are used throughout the class and book.
This is for quantitative researchers in the health, behavioral, or social sciences.
See: tysonstanley.github.io/Rstats for the book.
}
